Extract resolver env deployment support
